fix: añadi mas dependencias para eslint-prettier
