+++ src/components/App/App.component.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import './App.css';
+
+function App() {
+  return (
+    <div className="App">
+      <header className="App-header">
+        <p>hola</p>
+      </header>
+    </div>
+  );
+}
+
+export default App;
