+++ src/components/Content/Content.component.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function Content() {
+  return <div>Display content</div>;
+}
+
+export default Content;
